requirements.txt
fastapi
uvicorn
tortoise-orm
aiosqlite
jinja2
pillow
python-multipart
passlib
itsdangerous
argon2-cffi
cryptography
